Move and rename element classes
